refactor: extract pure helpers into Builder.Dispatch
Moves `DiagnosticMap`, `mergeDiagnostics`, `KnownTargetNames`,
`resolveKnownTargets`, `fileMatchesAnyTarget`, `pathSuffixesAsModuleName`,
and `filterToWatchDirs` out of `Tricorder.Builder` into a new
`Tricorder.Builder.Dispatch` module. `Builder.hs` is now ~100 lines
shorter and the effectful orchestration is no longer interleaved with
pure parsing and diagnostic merging.
